ERR: Hacked generated code to remove warnings.

+++ StreamedWatershedSegmentation/ParameterFileParser/y.tab.cxx
@@ -919,7 +919,7 @@ yyerrlab:   /* here on detecting error */
     count = 0;
     /* Start X at -yyn if nec to avoid negative indexes in yycheck.  */
     for (x = (yyn < 0 ? -yyn : 0);
-         x < (sizeof(yytname) / sizeof(char *)); x++)
+         x < static_cast<int>(sizeof(yytname) / sizeof(char *)); x++)
       if (yycheck[x + yyn] == x)
         size += strlen(yytname[x]) + 15, count++;
     msg = (char *) malloc(size + 15);
@@ -931,7 +931,7 @@ yyerrlab:   /* here on detecting error */
     {
       count = 0;
       for (x = (yyn < 0 ? -yyn : 0);
-           x < (sizeof(yytname) / sizeof(char *)); x++)
+           x < static_cast<int>(sizeof(yytname) / sizeof(char *)); x++)
         if (yycheck[x + yyn] == x)
           {
       strcat(msg, count == 0 ? ", expecting `" : " or `");
